pyproject.toml
[project]
name = "whisplaychat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.1.1",
    "pyaudio>=0.2.14",
    "pygame>=2.6.1",
    "python-dotenv>=1.2.2",
    "websockets>=16.0",
]
